+++ resolwe_bio/processes/workflows/bbduk_star_featurecounts_qc.py
@@ -46,7 +46,7 @@ class WorkflowBBDukStarFcQC(Process):
     entity = {
         "type": "sample",
     }
-    version = "6.2.0"
+    version = "6.3.0"
     process_type = "data:workflow:rnaseq:featurecounts:qc"
     category = "Pipeline"
 
@@ -763,25 +763,5 @@ class WorkflowBBDukStarFcQC(Process):
             ],
             "advanced": {"dirs": True, "config": True},
         }
-
-        # RNA-SeQC tool is initiated only if annotation source is ENSEMBL
-        if inputs.annotation.output.source == "ENSEMBL":
-            input_rnaseqc = {
-                "alignment": alignment_downsampled,
-                "annotation": inputs.annotation,
-                "strand_detection_options": {"stranded": inputs.assay_type},
-            }
-
-            if inputs.cdna_index:
-                input_rnaseqc["strand_detection_options"][
-                    "cdna_index"
-                ] = inputs.cdna_index
-
-            rnaseqc = Data.create(
-                process=BioProcess.get_latest(slug="rnaseqc-qc"),
-                input=input_rnaseqc,
-                name=f"RNA-SeQC QC report ({inputs.reads.name})",
-            )
-            input_multiqc["data"].append(rnaseqc)
 
         Data.create(process=BioProcess.get_latest(slug="multiqc"), input=input_multiqc)