+++ docs/CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
 Changed
 -------
 - Remove ``rnaseqc-qc`` 3' bias statistics from MultiQC report
+- Remove ``rnaseqc-qc`` from RNA-seq workflows
 
 Fixed
 -----